+++ app/src/main/java/jp/co/yumemi/droidtraining/MainViewModel.kt
@@ -3,7 +3,7 @@ package jp.co.yumemi.droidtraining
 import androidx.compose.runtime.Stable
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import jp.co.yumemi.repository.YumemiWeather
+import jp.co.yumemi.datasource.YumemiWeather
 import jp.co.yumemi.droidtraining.core.common.suspendRunCatching
 import jp.co.yumemi.droidtraining.core.model.Weather
 import kotlinx.coroutines.flow.MutableStateFlow
