+++ core/datasource/.gitignore
