+++ core/datasource/proguard-rules.pro
