+++ core/datasource/src/main/AndroidManifest.xml
