+++ core/datasource/src/main/java/jp/co/yumemi/datasource/model/Weather.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.repository.model
+package jp.co.yumemi.datasource.model
 
 internal enum class Weather {
     Sunny, Cloudy, Rainy, Snow
