+++ core/datasource/src/main/java/jp/co/yumemi/datasource/model/WeatherRequest.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.repository.model
+package jp.co.yumemi.datasource.model
 
 import com.squareup.moshi.JsonClass
 import java.util.Date
